+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - server: fix query execution of custom function containing a composite argument type
 - server: fix a bug in query validation that would cause some queries using default variable values to be rejected (fix #6867)
 - server: custom URI schemes are now supported in CORS config (fix #5818) (#5940)
+- server: fix MSSQL multiplexed subscriptions (fix #6887)
 - console: read-only modify page for mssql
 - console: filter out partitions from track table list and display partition info
 - console: fixes an issue where no schemas are listed on an MSSQL source